refactor: move Search page into TeacherViews
