+++ src/components/TeacherViews/SearchComponents/Search.view.test.tsx
@@ -2,11 +2,11 @@ import { waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { act } from "react-dom/test-utils";
 
-import renderWithProviders from "../../__tests__/__helpers__/renderWithProviders";
 
-import Search from "./Search.page";
-import { SearchProps } from "./search.page.types";
+import Search from "./Search.view";
+import { SearchProps } from "./search.view.types";
 
+import renderWithProviders from "@/__tests__/__helpers__/renderWithProviders";
 import { SearchHit, SearchQuery } from "@/context/Search/search.types";
 
 const createSearchResult = (): SearchHit => {
