+++ src/components/TeacherViews/SearchComponents/Search.view.tsx
@@ -1,11 +1,11 @@
 import { useRouter } from "next/router";
 import { FC, useEffect } from "react";
 
-import { SearchResultsItemProps } from "../SearchResultsItem/SearchResultsItem";
 
-import { SearchProps } from "./search.page.types";
+import { SearchProps } from "./search.view.types";
 import { isKeyStageTitleValueType, removeHTMLTags } from "./helpers";
 
+import { SearchResultsItemProps } from "@/components/SearchResultsItem/SearchResultsItem";
 import { Heading } from "@/components/SharedComponents/Typography";
 import useAnalytics from "@/context/Analytics/useAnalytics";
 import useAnalyticsPageProps from "@/hooks/useAnalyticsPageProps";
