+++ src/components/TeacherViews/SearchComponents/helpers/index.test.ts
