+++ src/components/TeacherViews/SearchComponents/helpers/index.ts
