+++ src/components/TeacherViews/SearchComponents/search.view.types.ts
